requirements.txt
Flask==0.10.1
python-firebase==1.2